calvinsys: split serverhandler and webserverhandler
For websocket handling to be migrated as expected a spilt was needed
